Atualizei a parte do cadastro quem quer ser ninja
